# reordered cli arguments for ffmpeg to improve speed to generate thumbnails of videos