--- a/api/scanner/media_encoding/executable_worker/executable_worker.go
+++ b/api/scanner/media_encoding/executable_worker/executable_worker.go
@@ -143,12 +143,12 @@ func (worker *FfmpegWorker) EncodeVideoThumbnail(inputPath string, outputPath st
 	thumbnailOffsetSeconds := fmt.Sprintf("%d", int(probeData.Format.DurationSeconds*0.25))
 
 	args := []string{
+		"-ss", thumbnailOffsetSeconds, // grab frame at time offset
 		"-i",
 		inputPath,
 		"-vframes", "1", // output one frame
 		"-an", // disable audio
 		"-vf", "scale='min(1024,iw)':'min(1024,ih)':force_original_aspect_ratio=decrease:force_divisible_by=2",
-		"-ss", thumbnailOffsetSeconds, // grab frame at time offset
 		outputPath,
 	}