Currently, Photoview uses output seeking of ffmpeg to generate thumbails of videos. This is a very time consuming behavior and not necessary since ffmpeg 2.1 (which is very outdated).

Reordering the arguments in the way this PR introduces changes the behavior to input seeking which (instead of encoding the full video) just seeks through the video based on key frames.

For a better and more in depth explanation of the behavior and the pros and cons see the official ffmpeg documentation (https://trac.ffmpeg.org/wiki/Seeking#Inputseeking).

Love to get feedback on this PR (I understand that the maintainer has paused the project, but I'd be happy to hear already what @kkovaletp thinks).