@kkovaletp 2024-03-15
Nice catch! As this is just a reordering of existing parameters, I don't expect any potential issues, introduced by the PR. Did you measure the difference in the processing time and the overall job run?