@AndGem 2024-03-15
I didn't measure it per se, but was comparing before and after behavior of my local server that is running photoview, but did a test for ffmpeg's performance:

Old ordering:
```
time ffmpeg -i ElephantsDream.mp4 -vframes 1 -an -vf "scale='min(1024,iw)':'min(1024,ih)':force_original_aspect_ratio=decrease:force_divisible_by=2" -ss 300 ./test2.jpg
```
```
46.87s user 0.14s system 203% cpu 23.093 total
```

New ordering:
```
time ffmpeg -ss 300 -i ElephantsDream.mp4 -vframes 1 -an -vf "scale='min(1024,iw)':'min(1024,ih)':force_original_aspect_ratio=decrease:force_divisible_by=2" ./test3.jpg
```
```
0.12s user 0.03s system 128% cpu 0.113 total
```

Obviously this is a non-scientific comparison, but the results convinced me.